test: switch to node:test
This commit switches to using node:test for the test runner.
