--- .eslintrc.js
@@ -1,20 +1,0 @@
-module.exports = {
-  env: {
-    commonjs: true,
-    es6: true,
-    node: true,
-  },
-  extends: [
-    'airbnb-base',
-  ],
-  globals: {
-    Atomics: 'readonly',
-    SharedArrayBuffer: 'readonly',
-  },
-  parserOptions: {
-    ecmaVersion: 2018,
-  },
-  rules: {
-    "key-spacing" : ["error", { "beforeColon" : true, "afterColon" : true}],
-  },
-};
