+++ package.json
@@ -6,7 +6,7 @@
   "author": "jniles",
   "license": "MIT",
   "scripts": {
-    "test": "ava",
+    "test": "node --test test/pdf.test.js",
     "release": "release-it --disable-metrics"
   },
   "publishConfig": {
@@ -26,15 +26,14 @@
   "dependencies": {
     "debug": "^4.1.1",
     "inline-source": "^8.0.2",
-    "puppeteer": "^23.4.0",
+    "puppeteer": "^24.2.0",
     "puppeteer-cluster": "^0.24.0"
   },
   "devDependencies": {
-    "ava": "^5.3.1",
-    "eslint": "^8.57.1",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.22.1",
+    "@eslint/js": "^9.19.0",
+    "eslint": "^9.19.0",
     "file-type": "^16.2.0",
-    "release-it": "^17.4.1"
+    "globals": "^15.14.0",
+    "release-it": "^18.1.2"
   }
 }
